refactor files, folders and variables
