+++ src/components/Home/HomeResources/HomeResources.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 
 import { Stack, Grid } from '@mui/material';
 
-import HomeRoomHeader from '@components/Home/HomeRooms/HomeRoomHeader/HomeResourceHeader';
-import HomeRoomList from '@components/Home/HomeRooms/HomeRoomList/HomeResourceList';
+import HomeRoomHeader from '@/components/Home/HomeResources/SearchFilterView/SearchFilterView';
+import HomeRoomList from '@/components/Home/HomeResources/HomeResourceList/HomeResourceList';
 import Room from '@/models/room';
 
 type Props = {
-  input: string;
+  searchedInput: string;
   rooms: Room[];
 };
 
@@ -15,7 +15,7 @@ const HomeRooms = (props: Props) => {
   return (
     <Grid container className='pl-10 md:justify-center md:pl-0'>
       <Stack spacing={-7} className='w-screen max-w-[1000px]'>
-        <HomeRoomHeader input={props.input} />
+        <HomeRoomHeader searchedInput={props.searchedInput} />
         <HomeRoomList rooms={props.rooms} />
       </Stack>
     </Grid>
